add test for n=92 (last value before long overflows)

+++ src/main/java/org/example/Fibonacci.java
@@ -1,19 +1,17 @@
 package org.example;
 
 public class Fibonacci {
-    public static long fib(int n) {
+    public static long  fib(int n) {
         if (n == 0) {
             return 0;
         }
-        int small = 0;
-        int large = 1;
+        long small = 0;
+        long large = 1;
         for (int i = 2; i <= n; i++) {
-            int temp = large;
+            long temp = large;
             large = small + large;
             small = temp;
         }
         return large;
     }
-
-
 }
